Move mag_ab and SpectrumTemplates to new photometry module
