+++ docs/photometry.rst
@@ -1,0 +1,18 @@
+*******************************
+Photometry (`skypy.photometry`)
+*******************************
+
+The following models are found in the `skypy.photometry` package.
+
+.. currentmodule:: skypy.photometry
+.. autosummary::
+   :nosignatures:
+
+   mag_ab
+   SpectrumTemplates
+
+
+Reference/API
+=============
+
+.. automodapi:: skypy.photometry
