+++ skypy/__init__.py
@@ -15,4 +15,5 @@ from ._astropy_init import *   # noqa
 __all__ = []
 
 from . import galaxy  # noqa
+from . import photometry  # noqa
 from . import pipeline  # noqa
