+++ skypy/photometry/__init__.py
@@ -1,0 +1,6 @@
+"""
+This module contains methods that model spectral energy distributions and
+calculate photometric properties.
+"""
+
+from ._photometry import *  # noqa F401,F403
